refactor(dwca): move catalog data with editor
