+++ specifyweb/frontend/js_src/package.json
@@ -17,7 +17,7 @@
   },
   "scripts": {
     "//": "All of these should be executed from the js_src directory",
-    "dwca:catalog": "node --import tsx lib/components/ExportFeed/data/generateGbifCatalog.ts",
+    "dwca:catalog": "node --import tsx lib/components/DwcaDefinition/data/generateGbifCatalog.ts",
     "localizationTests": "node --import ./lib/tests/jsdom.js --import ./lib/tests/register.js --import tsx lib/localization/__tests__/scanUsages.ts",
     "pullFromWeblate": "node --import ./lib/tests/jsdom.js --import ./lib/tests/register.js --import tsx lib/localization/utils/pull.ts",
     "schema:extract": "node --import ./lib/tests/jsdom.js --import ./lib/tests/register.js --import tsx lib/localization/schema-localization/extract.ts",
